[AArch64] Lower @llvm.ret.popless in swiftcorocc functions.
On AArch64, swiftcorocc functions are the only functions yet that can
support popless returns.

In the backend, that's done by recognizing the musttail call to
llvm.ret.popless preceding a ret instruction, and asking the target to
adjust that ret to be popless.

Throughout most of the backend, that's not an interesting difference.

In frame lowering, these popless rets now induce several special
behaviors in their (never shrink-wrapped) epilogues, all consequences
of not restoring SP:
- they of course don't do the SP adjustment or restore itself.
- most importantly, they force the epilogue callee-save restores
  to be FP-based rather than SP-based.
- they restore FP/LR last, as we still need the old FP, pointing
  at the frame being destroyed, to do the CSR restoring.
- with ptrauth-returns, they first derive the entry SP from
  FP, into X16, to use as a discriminator for a standalone AUTIB.

rdar://135984630