Define 'swiftcorocc' calling convention.
The 'swiftcorocc' calling convention is a variant of 'swiftcc',
but additionally allows the 'swiftcorocc' function to have
partial-pop/restore-less returns.

These restore-less returns don't fully restore the stack in their
epilog, thereby allowing the caller to access some stack allocations
made in the 'swiftcorocc' callee.

As part of that, these returns don't restore SP (but do restore FP).

So the most important characteristic of a 'swiftcorocc' call is that it
forces the caller function to access its stack through FP, like it does
with e.g., variable-size allocas.

This patch only implements the 'swiftcorocc' keyword and CallingConv,
but doesn't implement its support on any target yet.