[lldb] Change the type alias resolution algorithm
from a post-oder to a pre-order traversal.  This is necessary to
resolve generic type aliases that bind other type aliases in one go,
instead of first resolving the bound type aliases.  Debug Info will
have a record for `SomeAlias<SomeOtherAlias>` but not
`SomeAlias<WhatSomeOtherAliasResolvesTo>` because it tries to preserve
all sugar.

rdar://144884987