[lldb] Allow SwiftASTContext fallbacks for types from expressions
Generally, it would be feasible to rely on DWARF and reflectio
metadata for JIT-compiled images, but in practice, not everything that
an expression returns may be anchored by a variable, which makes
having debug info for expression-defined types hit and miss. If an
expression type is involved, LLDB will have done the costly
SwiftASTContext initialization already, so there isn't much saved by
avoiding the fallback in this case.