[lldb][formatter] Add children of Swift Tasks (#10057)
When displaying a Swift `Task`, include its child tasks.

This adds a `children` child to the `Task` synthetic provider. The type of `children` is a dynamically constructed tuple of N `UnsafeCurrentTask`, where N is the number of child tasks.

This allows users to see structured concurrency child tasks, in particular `async let` child tasks, and `TaskGroup` child tasks.