[InstrRef] Preserve debug instr num in aarch64-expand-pseudo LOADgot expansion (#128081)
The aarch64-expand-pseudo pass expands the LOADgot instruction to an
ADRP instruction and a LDRXui instruction. If the LOADgot had a
debug-instr-number, the pass doesn't preserve this to the new expansion.

This patch fixes the issue by making sure the debug-instr-number is
correctly applied to the LDRXui instruction generated.

(cherry picked from commit b3c51db292f05cf89201911cbcca6cba83caadd6)