[lldb] Improve handling of indirect enums using reflection metadata
This adds tests for several kinds of indirect enums that were
previously uncovered, such as GADTs and generic single-payload enums
and implements the proper handling of the ClosureContext box these
payloads come packaged in.

This fixes a lot of edge cases that previously required a
SwiftASTContext to resolve properly.

rdar://145301500