[clang dependency scanning] Make Current Working Directory Optimization Off by Default (#129809)
https://github.com/llvm/llvm-project/pull/124786 implemented current
working directory (CWD) optimization and the optimization was on by
default. We have discovered that build system needs to be compatible
with the CWD optimization and default off is a better behavior. The
build system needs to be aware that the current working directory is
ignored. Without a good way of notifying the build system, it is less
risky to default to off. This PR implement the change.

rdar://145860213
(cherry picked from commit 7bd492ffe84b549bc0f3b26e3aab973fa81069a5)

 Conflicts:
	clang/include/clang/Tooling/DependencyScanning/DependencyScanningService.h