[lldb] Accept address in task backtrace and select commands (#10191)
Update the `language swift task {backtrace,select}` commands to take a task address.

This is useful in conjunction with https://github.com/swiftlang/llvm-project/pull/10176 which prints a tasks address. An example use case is printing a `TaskGroup`, and then using the address of one of the group's child tasks with either the `backtrace` or `select` command.