[lldb][swift] Add function to extract funclet numbers
Async functions are broken into many pieces, and each piece is
identified by an integer number in the mangling scheme. This establishes
an ordering between funclets which we can exploit, in a subsequent
commit, to filter out multiple breakpoints locations attributed to the
same line.

An exception: entry funclets don't have such a number. This commit
assigns 0 to them, to better reflect the intent of ordering funclets.