[lldb][swift] Call FixCodeAddress on code addresses read from memory
In the changed code, we read a code address from memory and then attempt
to use that value to create a load address. This fails in targets using
high order bits for metadata, as we must first call `FixCodeAddress` so
that the appropriate plugin can clean-up the address.

Unfortunately, there doesn't seem to be any clean way of testing this,
but this patch also adds logging in case we fail to resolve the symbol
context again.