[lldb] Add initial formatter for Swift Actors (#10367)
Adds a synthetic formatter for `Builtin.DefaultActorStorage`. Each actor instance has a field named `$defaultActor` which has a type of `DefaultActorStorage`. Prior to this, that field was shown as opaque bytes. With this change, the value now has a single child, `unprioritised_jobs`, which represents the job queue for the actor.

This is the first of a few changes, including a summary formatter which will show which state the actor is in (ie "running", "idle", etc).