[lldb] SwiftUserExpression avoid call to ImportType (NFCi)
Currently SwiftASTContext::ImportType was is returning a
TypeSystemSwiftForExpressions TypeRef if the input is a TypeRef
type. This makes future type alias resolution very slow because we
lose the contextual information. By keeping it in the
TypeSystemSwiftTypeRef returned by the Variable these lookups remain
local. This can elminiate up to 90% of the time spent evaluating
expressions in extreme cases.

rdar://145884579