[lldb] Exclude redundant task flags from summary (#10374)
Don't show `future` or `childTask` flags in a Task's summary.

First, `future` is not an official term in Swift Concurrency. There's undetermined value
in showing this. From my reading of the concurrency runtime, in non-embedded Swift all
tasks have the flag `isFuture`. For embedded Swift it's _unset_ for all tasks.

Second, child tasks are already indicated with either `groupChildTask`, or
`asyncLetTask`.

Both values are still available on demand, ex `p task.isFuture` or `p task.isChildTask`.

This change is to reduce the quantity of data in the summary, because a subsequent
change will add the Task's name in its summary.