[lldb][NFC] Reduce scope of Swift customizations
The code in Process::HandleProcessStateChangedEvent is pretty
complicated, and a lot of Swift customizations were added on top of it
to support the repl. This commit reduces the area of those
customizations by moving them into a separate function; in particular,
the code computing whether to "check for a repl breakpoint" was factored
out.

The main motivation is that we would like to make some changes in
HandleProcessStateChangedEvent, which would run into nasty merge
conflicts otherwise.