[lldb] Temporarily disable actor state test (#10451)
* [lldb] Temporarily disable actor state test

The test added in "Add summary formatter for DefaultActorStorage (#10388)"
is checking the actor's state, but it is failing sometimes.
Temporarily disable until Dave can look into this, to unblock
CI/PR testing.

rdar://148754221

* Revert "[lldb] Temporarily disable actor state test"

This reverts commit 937c1859be608b8cdb4cfbee9140f47b45e01bd8.

* [lldb] Temporarily disable actor state test

The test added in "Add summary formatter for DefaultActorStorage (#10388)"
is checking the actor's state, but it is failing sometimes.
Temporarily disable until Dave can look into this, to unblock
CI/PR testing.

rdar://148754221