[InlineCost] Simplify extractvalue across the callsite
Motivation: When using libc++, `std::bitset<64>::count()` didn't
optimize to a single popcount instruction on AArch64, because we failed
to inline the library code completely. Inlining fails, because the
internal bit_iterator struct is passed as a [2 x i64] %arg value on
AArch64, built using insertvalue instructions. One of the array entries
is constant and if we know that this entry is constant, we can prove
that half the function becomes dead. However, InlineCost only considers
operands for simplification if they are Constants, which %arg is not.
Without this simplification the function is too expensive to inline.

Therefore, we need to teach InlineCost to simplify values which aren't
constants and which may originate from the caller function.

As a first step, we enable this for extractvalue, so it can be
simplified against insertvalues from the caller function.

There are similar opportunities we can explore for BinOps in the future
(e.g. cmp eq %arg1, %arg2 when the caller passes the same value into
both arguments), but we need to be careful here, because InstSimplify
isn't completely safe to use with instructions owned by another
function.